solve problem of price and decimal

+++ db/migrate/20201201121952_create_orders.rb
@@ -3,7 +3,7 @@ class CreateOrders < ActiveRecord::Migration[5.2]
     create_table :orders do |t|
       t.string :stripe_customer_id
       t.references :user, null: false, foreign_key: true
-      t.decimal       :total_price
+      t.integer    :total_price
       t.timestamps
     end
   end
